Feedback: reduce and rename
