+++ src/NuGet.Clients/NuGet.Tools/Resources.resx
@@ -198,6 +198,7 @@
   </data>
   <data name="Text_ConfigurationFiles_CommonLink" xml:space="preserve">
     <value>Common NuGet configurations: [How settings are applied](https://aka.ms/nuget/how-settings-are-applied/)</value>
+    <comment>Please leave the [ and ] brackets in place and do translate the words between them. Do not translate the URL inside ( ).</comment>
   </data>
   <data name="Text_FilePath_Header" xml:space="preserve">
     <value>File Path</value>
